+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add keyset pagination support via `pagination="keyset"` for better performance on large datasets (#15)
 * Allow filters for non-existent fields when a custom `filter_fn` is attached
+* Add basic `autocomplete` filter type, for searchable dropdowns with large preloaded option lists
 
 ## v0.9.0-beta.1 (2025-12-15)